feat: added support for headers when calling ovh.Client.raw_call
This feature aims at making available the use of headers while calling the OVH API.
The original behavior is preserved, as the signature changes with the introduction
of a non-mandatory parameter "header", which is None by default.
Added a piece of documentation to warn the user about the possbility
to have some headers being rewritten by the Client engine.

Signed-off-by: Geoffrey Bauduin <geoffrey.bauduin@corp.ovh.com>